Updated decorators

+++ todo/views.py
@@ -5,6 +5,7 @@ from django.http import HttpResponseRedirect
 from django.urls import reverse
 from django.shortcuts import get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 
 # Create your views here.
 
@@ -16,8 +17,9 @@ def get_showing_todos(request, todos):
         if request.GET.get('filter')=='incomplete':
             return todos.filter(is_completed=False)
     
-    return todos    
-
+    return todos
+    
+@login_required
 def index(request):
     todos = Todo.objects.filter(owner=request.user)
 
@@ -29,7 +31,7 @@ def index(request):
                 'completed_count': completed_count, 'incompleted_count': incompleted_count}
     return render(request, 'todo/index.html', context)
 
-
+@login_required
 def create_todo(request):
     form = TodoForm()
     context = {'form':form}
@@ -45,7 +47,9 @@ def create_todo(request):
         todo.is_completed = True if is_completed == 'on' else False
         todo.owner = request.user
 
-        todo.save()
+        if todo.owner== request.user:
+
+            todo.save()
         messages.add_message(request, messages.SUCCESS, "Todo created successfully")
 
         return HttpResponseRedirect(reverse("todo", kwargs={'id': todo.pk}))
@@ -53,23 +57,29 @@ def create_todo(request):
 
 
     return render(request, 'todo/create-todo.html', context)
-
+@login_required
 def todo_detail(request, id):
     todo = get_object_or_404(Todo, pk=id)
     context = {'todo':todo}
     return render(request, 'todo/todo-detail.html', context)
 
+@login_required
 def todo_delete(request, id):
     todo = get_object_or_404(Todo, pk=id)
     context = {'todo':todo}
+
     if request.method =="POST":
-        todo.delete()
 
-        messages.add_message(request, messages.SUCCESS, "Todo Deleted successfully")
-        return HttpResponseRedirect(reverse("home"))
+        if todo.owner== request.user:
+
+            todo.delete()
+
+            messages.add_message(request, messages.SUCCESS, "Todo Deleted successfully")
+            return HttpResponseRedirect(reverse("home"))
+        return render(request, 'todo/todo-delete.html', context)
         
-    return render(request, 'todo/todo-delete.html', context)
-
+        
+@login_required
 def todo_edit(request, id):
     todo = get_object_or_404(Todo, pk=id)
     form = TodoForm(instance=todo)
